add heartbeat producer and consumer examples
These examples demonstrate usage of NMT heartbeat protocol.

The two examples can be executed on different devices on the same network.
The consumer monitors the bus for heartbeats from the producer every
five seconds. Relevant configuration is done in .xpd profile files and
exported as OD.c and OD.h.